aider: Specify jinja2 version 2.11.3 in requirements.txt.
# Aider chat conversation:

ASSISTANT: The error message indicates that the `escape` function cannot be imported from the `jinja2` package. This could be due to an incorrect version of `jinja2` installed in your environment.

The `requirements.txt` file shows that `jinja2` is not versioned. This means that the latest version of `jinja2` is installed when you run `pip install -r requirements.txt`. If the latest version of `jinja2` does not have the `escape` function, this could cause the import error.

To fix this issue, you can specify a version of `jinja2` that includes the `escape` function in the `requirements.txt` file.

Here is the *SEARCH/REPLACE block* to modify the `requirements.txt` file:

```python
requirements.txt
<<<<<<< SEARCH
jinja2
=======
jinja2==2.11.3
>>>>>>> REPLACE
```

After making this change, you should reinstall the requirements using `pip install -r requirements.txt` and then try running the tests again.