aider: Set `extra_files` parameter of `app.run` to include the path to the secrets file, enabling Flask's reloader to restart the server when the secrets file changes.
# Aider chat conversation:

USER: Make login_app.py reload when changes are made to files including the secrets file.

ASSISTANT: To make the Flask application in `login_app.py` reload when changes are made to files, we can set the `debug` mode to `True` when running the application. This will enable the reloader, which watches for changes in your application and restarts the server.

However, Flask's reloader does not watch for changes in all files by default. It only watches Python files and some other types of files that are likely to be used in a Flask application. To make it watch the secrets file, we need to add the secrets file to the list of extra files that the reloader should watch.

Here is the *SEARCH/REPLACE block* for the change:

```python
src/login_app.py
<<<<<<< SEARCH
if __name__ == '__main__':
    app.run(debug=True)
=======
if __name__ == '__main__':
    app.run(debug=True, extra_files=['./secrets'])
>>>>>>> REPLACE
```

This change sets the `extra_files` parameter of the `run` method to a list containing the path to the secrets file. This tells Flask's reloader to restart the server when the secrets file changes.