aider: Fixed the issue with the `Jinja2Templates` class not finding the templates directory by using the absolute path of the `templates` directory relative to the `login.py` file.
# Aider chat conversation:

USER: I am still seeing 404's and no debugging to console/uvicorn

ASSISTANT: It seems like the issue might be related to the path where the `login.html` template is located. The FastAPI `Jinja2Templates` class expects the templates to be in a directory relative to the root directory of the project.

Currently, the `Jinja2Templates` is initialized with the directory set to `"templates"`, which means it will look for the templates in a directory named `templates` in the same directory where the `uvicorn` server is started. If the server is started from a different directory, it might not find the `templates` directory.

To fix this, we can use the `os` module to get the absolute path of the `templates` directory relative to the `login.py` file. This way, no matter from where the server is started, it will always find the `templates` directory.

Here is the *SEARCH/REPLACE block* for the change:

```python
src/login.py
<<<<<<< SEARCH
templates = Jinja2Templates(directory="templates")
=======
templates_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), "templates")
templates = Jinja2Templates(directory=templates_dir)
>>>>>>> REPLACE
```

This change will ensure that the `Jinja2Templates` class always looks for templates in the correct directory.

Please apply this change and let me know if you need further assistance.