aider: Changed path to the `templates` directory in `src/main.py` to point to the correct location.
# Aider chat conversation:

USER: login.html is in the templates directory at the root of this project. \

ASSISTANT: I see. The `login.html` file is located in the `templates` directory at the root of the project. However, the `Jinja2Templates` instance in `src/main.py` is looking for templates in a `templates` directory relative to the `src` directory.

To fix this, we need to change the path to the `templates` directory in `src/main.py` to point to the correct location.

Here is the *SEARCH/REPLACE block* for the change:

```python
src/main.py
<<<<<<< SEARCH
# Create Jinja2Templates instance
templates = Jinja2Templates("../templates")
=======
# Create Jinja2Templates instance
templates = Jinja2Templates("./templates")
>>>>>>> REPLACE
```

This change will make the `Jinja2Templates` instance look for templates in the `templates` directory at the root of the project, which is where the `login.html` file is located.